fixed more cast issues

+++ kuzu/src/main/java/org/ldbcouncil/snb/impls/workloads/kuzu/KuzuDb.java
@@ -171,9 +171,9 @@ public class KuzuDb extends BaseDb<KuzuQueryStore>
             long personId = record.getValue( 0 ).getValue();
             String personFirstName = record.getValue( 1 ).getValue();
             String personLastName = record.getValue( 2 ).getValue();
-            int xCount = record.getValue( 3 ).getValue();
-            int yCount = record.getValue( 4 ).getValue();
-            int count = record.getValue( 5 ).getValue();
+            int xCount = (int)record.getValue( 3 ).getValue();
+            int yCount = (int)record.getValue( 4 ).getValue();
+            int count = (int)record.getValue( 5 ).getValue();
             return new LdbcQuery3Result(
                     personId,
                     personFirstName,
@@ -200,7 +200,7 @@ public class KuzuDb extends BaseDb<KuzuQueryStore>
         public LdbcQuery4Result toResult( KuzuFlatTuple record ) throws KuzuObjectRefDestroyedException
         {
             String tagName = record.getValue( 0 ).getValue();
-            int postCount = record.getValue( 1 ).getValue();
+            int postCount = (int)record.getValue( 1 ).getValue();
             return new LdbcQuery4Result( tagName, postCount );
         }
     }
@@ -221,7 +221,7 @@ public class KuzuDb extends BaseDb<KuzuQueryStore>
         public LdbcQuery5Result toResult( KuzuFlatTuple record ) throws KuzuObjectRefDestroyedException
         {
             String forumTitle = record.getValue( 0 ).getValue();
-            int postCount = record.getValue( 1 ).getValue();
+            int postCount = (int)record.getValue( 1 ).getValue();
             return new LdbcQuery5Result( forumTitle, postCount );
         }
     }
@@ -242,7 +242,7 @@ public class KuzuDb extends BaseDb<KuzuQueryStore>
         public LdbcQuery6Result toResult( KuzuFlatTuple record ) throws KuzuObjectRefDestroyedException
         {
             String tagName = record.getValue( 0 ).getValue();
-            int postCount = record.getValue( 1 ).getValue();
+            int postCount = (int)record.getValue( 1 ).getValue();
             return new LdbcQuery6Result( tagName, postCount );
         }
     }
@@ -268,7 +268,7 @@ public class KuzuDb extends BaseDb<KuzuQueryStore>
             long likeCreationDate = record.getValue( 3 ).getValue();
             long messageId = record.getValue( 4 ).getValue();
             String messageContent = record.getValue( 5 ).getValue();
-            int minutesLatency = record.getValue( 6 ).getValue();
+            int minutesLatency = (int)record.getValue( 6 ).getValue();
             boolean isNew = record.getValue( 7 ).getValue();
             return new LdbcQuery7Result(
                     personId,
@@ -362,7 +362,7 @@ public class KuzuDb extends BaseDb<KuzuQueryStore>
             long personId = record.getValue( 0 ).getValue();
             String personFirstName = record.getValue( 1 ).getValue();
             String personLastName = record.getValue( 2 ).getValue();
-            int commonInterestScore = record.getValue( 3 ).getValue();
+            int commonInterestScore = (int)record.getValue( 3 ).getValue();
             String personGender = record.getValue( 4 ).getValue();
             String personCityName = record.getValue( 5 ).getValue();
             return new LdbcQuery10Result(
@@ -394,7 +394,7 @@ public class KuzuDb extends BaseDb<KuzuQueryStore>
             String personFirstName = record.getValue( 1 ).getValue();
             String personLastName = record.getValue( 2 ).getValue();
             String organizationName = record.getValue( 3 ).getValue();
-            int organizationWorkFromYear = record.getValue( 4 ).getValue();
+            int organizationWorkFromYear = (int)record.getValue( 4 ).getValue();
             return new LdbcQuery11Result(
                     personId,
                     personFirstName,
